requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
ultralytics
fastapi
uvicorn
python-multipart
jinja2
pillow